edit

+++ lab1_3.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main(){
-    unsigned long long int n,a,b;
+    unsigned long int n,a,b;
 
-    scanf("%lld",&n);
+    scanf("%u",&n);
     a=n%3;
     b=n%11;
     printf("%d %d",a,b);
